+++ tests/FileEntrySecretTests.cs
@@ -73,6 +73,33 @@ namespace Tests
 		}
 
 		[Test]
+		public void GetFileEntryTest()
+		{
+			// Arrange
+			byte[] derivedKey = new byte[16] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16 };
+			byte[] initialCounter = new byte[] { 0xf0, 0xf1, 0xf2, 0xf3, 0xf4, 0xf5, 0xf6, 0xf7, 0xf8, 0xf9, 0xfa, 0xfb, 0xfc, 0xfd, 0xfe, 0xff };
+
+			SettingsAES_CTR settingsAES_CTR = new SettingsAES_CTR(initialCounter);
+
+			SymmetricKeyAlgorithm skaAES_CTR = new SymmetricKeyAlgorithm(SymmetricEncryptionAlgorithm.AES_CTR, 128, settingsAES_CTR);
+
+			string filename = "ni12ce.pdf";
+			byte[] fileContent = new byte[] { 1, 2, 32, 11, 2, byte.MaxValue, 0, 0, 2, 34, 45, 0, 0, 0, 0};
+
+			FileEntry fe = new FileEntry(filename, fileContent);
+
+			FileEntrySecret fes = new FileEntrySecret(fe, "does not matter", skaAES_CTR, derivedKey);
+
+			// Act
+			FileEntry feCopy = fes.GetFileEntry(derivedKey);
+
+			// Assert
+			Assert.IsTrue(ComparisonHelper.AreFileEntriesEqual(fe, feCopy));
+			Assert.AreEqual(fe.creationTime, feCopy.creationTime);
+			Assert.AreEqual(fe.modificationTime, feCopy.modificationTime);
+		}
+
+		[Test]
 		public void GetFilenameTest()
 		{
 			// Arrange